+++ CHANGELOG.md
@@ -1,3 +1,4 @@
-# Version 26.8.29.3
+# Version 26.8.30.1
 
-- Trying a new style for web interface that does not contain a bounding box (making it similar to how claude chatbox is)
+- Added a camera input for perla, take picture and directly attach it to perla
+- Add up to 6 images in one messages